flask_news/requirements.txt
flask
flask-cors
apscheduler
sqlalchemy
python-dotenv
playwright
beautifulsoup4
openai
chromadb
uvicorn
requests
pymongo
firebase-admin
